Metrics: Move creation to the manager class
